#15: Fix command will ignore test versions
Update the code finding the right tag to base a fix version off to ignore
test versions and only check for actual versions (e.g. 'v0.1.0' is valid,
'v0.1.0-4-g93a2477' is not valid).
Update the output messages when creating fix branches to specify the source
tag to avoid confusion.