fix: support multiple digits in version string (#17)
Add a new method to run `git describe` with a match and an exclude pattern.
Update the test release worker to match any version tag in the format of
`v*.*.*` that does not include any `-` characters. This allows the pattern
to work for version numbers > 10 while still excluding previous test
versions (which always include a `-`).