feat: new checkout command (#20)
Add a new `checkout` command to allow checking out the tag for a specific
version. This re-uses functionality from the `fix` command.
Add new functions to find the git based on the version string provided by
the user. Rename the existing method that finds the latest bugfix based on
a given version to help avoid confusion.
Update project README with information on the new command.